+++ packages/client-auth/src/lib/config.ts
@@ -59,7 +59,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
     language: "typescript",
     openapiDocVersion: "0.0.0",
-    sdkVersion: "1.0.0-alpha.11",
-    genVersion: "2.378.1",
-    userAgent: "speakeasy-sdk/typescript 1.0.0-alpha.11 2.378.1 0.0.0 @formance/sdk-auth",
+    sdkVersion: "1.0.0-alpha.12",
+    genVersion: "2.378.3",
+    userAgent: "speakeasy-sdk/typescript 1.0.0-alpha.12 2.378.3 0.0.0 @formance/sdk-auth",
 } as const;